+++ .vscode/numbered-bookmarks.json
@@ -1,0 +1,3 @@
+{
+	"bookmarks": []
+}
