+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "start": "webpack-dev-server --config ./webpack.config.js --mode development"
   },
   "keywords": [],
@@ -17,6 +17,8 @@
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
     "css-loader": "^2.1.1",
+    "enzyme": "^3.10.0",
+    "jest": "^24.9.0",
     "react-hot-loader": "^4.12.8",
     "style-loader": "^0.23.1",
     "webpack": "^4.36.1",
@@ -24,6 +26,8 @@
     "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
+    "enzyme-adapter-react-15": "^1.4.1",
+    "enzyme-adapter-react-16": "^1.15.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6"
   }
