+++ src/App.js
@@ -1,0 +1,214 @@
+import React from 'react';
+// global variable for audio
+let sound;
+
+// React component
+export default class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      sessionTime: 25,
+      breakTime: 5,
+      currentTime: 1500,
+      running: 'Session',
+      paused: true,
+      countDown: ''
+    }
+  }
+
+  showTime = () => {
+    let mins = Math.floor(this.state.currentTime / 60);
+    let secs = this.state.currentTime - mins * 60;
+    mins = mins < 10 ? '0' + mins : mins;
+    secs = secs < 10 ? '0' + secs : secs;
+    return mins + ':' + secs;
+  }
+
+  handleCountdown = () => {
+    if (this.state.paused) {
+      this.startCountdown();
+      this.setState({
+        paused: false
+      });
+    } else {
+      this.state.countDown && clearInterval(this.state.countDown);
+      this.setState({
+        paused: true
+      });
+    };
+  }
+
+  timerSwitch = () => {
+    this.state.running === 'Session' ? (
+      this.setState({
+        currentTime: this.state.breakTime * 60,
+        running: 'Break'
+      })
+    ) : (
+      this.setState({
+        currentTime: this.state.sessionTime * 60,
+        running: 'Session'
+      })
+    )
+  }
+
+  startCountdown = () => {
+    this.setState({
+      countDown: setInterval(() => {
+        this.setState({
+          currentTime: this.state.currentTime - 1
+        });
+        this.playSound();
+        if (this.state.currentTime < 0) {
+          this.state.countDown && clearInterval(this.state.countDown),
+          this.timerSwitch(),
+          this.startCountdown()
+        };
+      }, 1000),
+    });
+  }
+
+  clearAll = () => {
+    this.state.countDown && clearInterval(this.state.countDown);
+    this.setState({
+      sessionTime: 25,
+      breakTime: 5,
+      currentTime: 1500,
+      running: 'Session',
+      paused: true
+    });
+    sound = document.getElementById('beep');
+    sound.currentTime = 0;
+    sound.pause();
+  }
+
+  sessionChange = e => {
+    this.lengthChange('sessionTime', e.target.getAttribute('value'), this.state.sessionTime, 'Session');
+  }
+
+  breakChange = e => {
+    this.lengthChange('breakTime', e.target.getAttribute('value'), this.state.breakTime, 'Break');
+  }
+
+  lengthChange = (targetState, val, length, timerType) => {
+    if (!this.state.paused) return;
+    if (this.state.running != timerType) {
+      if (val == '-' && length > 1) {
+        this.setState({
+          [targetState]: length - 1
+        });
+      } else if (val == '+' && length < 60) {
+        this.setState({
+          [targetState]: length + 1
+        });
+      };
+    } else {
+      if (val == '-' && length > 1) {
+        this.setState({
+          [targetState]: length - 1,
+          currentTime: length * 60 - 60
+        });
+      } else if (val == '+' && length < 60) {
+        this.setState({
+          [targetState]: length + 1,
+          currentTime: length * 60 + 60
+        });
+      }
+    };
+  }
+
+  playSound = () => {
+    if (this.state.currentTime === 0) {
+      sound = document.getElementById('beep');
+      sound.currentTime = 0;
+      sound.play();
+    };
+  }
+
+  render() {
+    return(
+      <div id='app'>
+        <h1>Pomodoro Clock</h1>
+        <div id='settings'>
+          <div id='session-label'>
+            <h3>Session Length</h3>
+            <div className='timeControls'>
+              <span>
+                <i className="fas fa-arrow-down"
+                   id='session-decrement'
+                   value='-'
+                   onClick={this.sessionChange}/>
+              </span>
+              <input
+                id='session-length'
+                className="number-input"
+                onChange={e => this.setState({
+                      sessionTime: e.target.value,
+                      currentTime: e.target.value * 60 + 60
+                   })}
+                value={this.state.sessionTime}
+              />
+              <span>
+                <i className="fas fa-arrow-up"
+                   id='session-increment'
+                   value='+'
+                   onClick={this.sessionChange}/>
+              </span>
+            </div>
+          </div>
+          <div id='break-label'>
+            <h3>Break Length</h3>
+            <div className='timeControls'>
+              <span>
+                <i className="fas fa-arrow-down"
+                   id='break-decrement'
+                   value='-'
+                   onClick={this.breakChange}/>
+              </span>
+              <input
+                id='break-length'
+                className="number-input"
+                value={this.state.breakTime}
+                onChange={e => this.setState({
+                    breakTime: e.target.value,
+                  })}
+              />
+              <span>
+                <i className="fas fa-arrow-up"
+                   id='break-increment'
+                   value='+'
+                   onClick={this.breakChange}/>
+              </span>
+            </div>
+          </div>
+        </div>
+        <div id='timer'>
+          <div id='time-left'>
+            <div id='timer-label'>
+            {this.state.running}
+              <audio
+                id='beep'
+                src='https://s3.amazonaws.com/freecodecamp/drums/Chord_3.mp3'
+                type='audio/mp3'
+                preload='auto'
+              />
+            </div>
+          {this.showTime()}
+          </div>
+        </div>
+        <div id='controls'>
+          <span id='start_stop' onClick={this.handleCountdown}>
+            {
+              this.state.paused
+              ? <i className="fas fa-play fa-2x"></i>
+              : <i className="fas fa-pause fa-2x"></i>
+            }
+          </span>
+          <span id='reset' onClick={this.clearAll}>
+            <i className="fas fa-redo-alt fa-2x"></i>
+          </span>
+        </div>
+      </div>
+    )
+  }
+}
